Add integration tests to increase stability assurance
Our current unit tests don't really give any assurance that git-clean
actually works, just that a few of the outputs remain the same. The only
way to really test that I haven't broken everything is through
integration tests, where we set up a git repo, commit some branches, and
try deleting them. This adds the integration tests framework, and some
of the core tests around deleting functionality, all using local
branches at the moment.

A git remote _can_ be a local folder, so that'll be the next target for
tests.