Squashes (#26)
* Make squashes an option, and assume all non-fastforwardable branches are
squashed

I don't know if this is a recent change in `git`, or if this just never
worked, but `git diff` is no longer a reliable way to check for a
squashed branch. Once the base branch gets ahead of the squashed branch,
the diff will always return something, making it practically impossible
to tell.

So instead, let's just make checking for squashed branches an option! If
a user chooses to opt in to this behavior, we'll consider every
non-fastforwardable branch to be squashed, and thus offer it up for
deletion. This workflow probably won't work for everyone.

If a user chooses not to opt in, great! They still get some branches
deleted, and it'll run a lot faster.

* Bump to version 0.5.0

* Fix tests on Rust 1.15

In the associated Cargo release, looks like a /target/debug/deps
directory was introduced, which is where env::current_exe thinks the
tests are. Instead, we can more reliably use the CARGO_MANIFEST_DIR
environment variable to determine the path to our executable